requirements.txt
# Core Orchestration Library
langchain>=0.1.0
langgraph>=0.0.30

# Web User Interface
gradio>=4.0.0

# Image Processing and Numerical Operations
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Type Hinting Compatibility
typing-extensions>=4.8.0